HoTT: refactoring: use more appropriate types
 * Use `bool`s instead of `int`s
 * Use `size_t` to hold buffer sizes, not `int`
 * Use `const` where data isn't or shouldn't be modified
  - send_data: don't modify the place that holds the checksum before
	       sending it, instead just send the checksum instead of
               that place.
 * Make objects read from multiple threads volatile
 * Don't rely on endianness
  - ... for reading into the correct part of an int when we're really
    reading an 8-bit (unsigned) integer: use `uint8_t` instead
 * Use `sizeof` to determine buffer sizes
  - don't hard code it as a magic number
 * When we want to read an *unsigned* (8-bit) integer *don't* use a `char`
  - `char` can be either signed or unsigned, use `uint8_t` instead where
    signdness matters
 * Use (new macro) `ARRAY_SIZE` to determine the size of an array
   instead of hardcoding it using a magic number
 * Don't "guess" a good buffer-size, when `sizeof` will do just fine
  - MESSAGE_BUFFER_SIZE is now #defined to a `sizeof` of a union of
    possible HoTT messages, this should *always* yield the correct
    buffer size

+++ apps/hott_telemetry/messages.c
@@ -54,7 +54,7 @@ void messages_init(void)
 	sensor_sub = orb_subscribe(ORB_ID(sensor_combined));
 }
 
-void build_eam_response(uint8_t *buffer, int *size)
+void build_eam_response(uint8_t *buffer, size_t *size)
 {
 	/* get a local copy of the current sensor values */
 	struct sensor_combined_s raw = { 0 };
